requirements.txt
PySide6
pyqtgraph
pyserial
requests
